requirements.txt
streamlit>=1.24.0
gitpython>=3.1.30
requests>=2.28.0
numpy>=1.21.0
plotly>=5.10.0
networkx>=3.0
matplotlib>=3.7.0
PyOpenGL>=3.1.6
PyOpenGL-accelerate>=3.1.6
